requirements.txt
annotated-types @ file:///home/conda/feedstock_root/build_artifacts/annotated-types_1733247046149/work
anyio @ file:///home/conda/feedstock_root/build_artifacts/anyio_1736174388474/work
asgiref==3.8.1
attrs==25.3.0
azure-ai-ml==1.26.0
azure-ai-textanalytics==5.3.0
azure-common==1.1.28
azure-core==1.32.0
azure-core-tracing-opentelemetry==1.0.0b11
azure-identity==1.21.0
azure-mgmt-core==1.5.0
azure-monitor-opentelemetry==1.6.5
azure-monitor-opentelemetry-exporter==1.0.0b35
azure-storage-blob==12.25.0
azure-storage-file-datalake==12.19.0
azure-storage-file-share==12.21.0
blinker @ file:///home/conda/feedstock_root/build_artifacts/blinker_1731096409132/work
certifi @ file:///home/conda/feedstock_root/build_artifacts/certifi_1739515848642/work/certifi
cffi==1.17.1
charset-normalizer==3.4.1
click @ file:///D:/bld/click_1734858834134/work
colorama @ file:///home/conda/feedstock_root/build_artifacts/colorama_1733218098505/work
cryptography==44.0.2
Deprecated==1.2.18
distro @ file:///home/conda/feedstock_root/build_artifacts/distro_1734729835256/work
exceptiongroup @ file:///home/conda/feedstock_root/build_artifacts/exceptiongroup_1733208806608/work
fixedint==0.1.6
Flask @ file:///home/conda/feedstock_root/build_artifacts/flask_1741793020411/work
h11 @ file:///home/conda/feedstock_root/build_artifacts/h11_1733327467879/work
h2 @ file:///home/conda/feedstock_root/build_artifacts/h2_1738578511449/work
hpack @ file:///home/conda/feedstock_root/build_artifacts/hpack_1737618293087/work
httpcore @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_httpcore_1731707562/work
httpx @ file:///home/conda/feedstock_root/build_artifacts/httpx_1733663348460/work
hyperframe @ file:///home/conda/feedstock_root/build_artifacts/hyperframe_1737618333194/work
idna @ file:///home/conda/feedstock_root/build_artifacts/idna_1733211830134/work
importlib_metadata @ file:///home/conda/feedstock_root/build_artifacts/importlib-metadata_1737420181517/work
isodate==0.7.2
itsdangerous @ file:///home/conda/feedstock_root/build_artifacts/itsdangerous_1733308265247/work
Jinja2 @ file:///home/conda/feedstock_root/build_artifacts/jinja2_1741263328855/work
jiter @ file:///D:/bld/jiter_1741655103938/work
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
MarkupSafe @ file:///D:/bld/markupsafe_1733219622495/work
marshmallow==3.26.1
msal==1.32.0
msal-extensions==1.3.0
msrest==0.7.1
numpy @ file:///D:/bld/numpy_1739425584338/work/dist/numpy-2.2.3-cp313-cp313-win_amd64.whl#sha256=d33e937ed2249dc8903e4afd8898b785eee9cde7e2f517ff584cbd7390f83579
oauthlib==3.2.2
openai @ file:///home/conda/feedstock_root/build_artifacts/openai_1741872554998/work
opentelemetry-api==1.31.0
opentelemetry-instrumentation==0.52b0
opentelemetry-instrumentation-asgi==0.52b0
opentelemetry-instrumentation-dbapi==0.52b0
opentelemetry-instrumentation-django==0.52b0
opentelemetry-instrumentation-fastapi==0.52b0
opentelemetry-instrumentation-flask==0.52b0
opentelemetry-instrumentation-psycopg2==0.52b0
opentelemetry-instrumentation-requests==0.52b0
opentelemetry-instrumentation-urllib==0.52b0
opentelemetry-instrumentation-urllib3==0.52b0
opentelemetry-instrumentation-wsgi==0.52b0
opentelemetry-resource-detector-azure==0.1.5
opentelemetry-sdk==1.31.0
opentelemetry-semantic-conventions==0.52b0
opentelemetry-util-http==0.52b0
packaging==24.2
pandas @ file:///D:/bld/pandas_1726878597308/work
psutil==6.1.1
pycparser==2.22
pydantic @ file:///home/conda/feedstock_root/build_artifacts/pydantic_1732286392017/work
pydantic_core @ file:///C:/b/abs_bdosz7qwys/croot/pydantic-core_1734726071532/work
pydash==8.0.5
PyJWT==2.10.1
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/python-dateutil_1733215673016/work
pytz @ file:///home/conda/feedstock_root/build_artifacts/pytz_1706886791323/work
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
requests-oauthlib==2.0.0
rpds-py==0.23.1
six @ file:///home/conda/feedstock_root/build_artifacts/six_1733380938961/work
sniffio @ file:///home/conda/feedstock_root/build_artifacts/sniffio_1733244044561/work
strictyaml==1.7.3
tqdm @ file:///home/conda/feedstock_root/build_artifacts/tqdm_1735661334605/work
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/typing_extensions_1733188668063/work
tzdata @ file:///home/conda/feedstock_root/build_artifacts/python-tzdata_1737541069190/work
urllib3==2.3.0
Werkzeug @ file:///home/conda/feedstock_root/build_artifacts/werkzeug_1733160440960/work
wrapt==1.17.2
zipp @ file:///home/conda/feedstock_root/build_artifacts/zipp_1732827521216/work
